Intensity dependent gn code is now written, and is ready for proper testing.

+++ src/histogram/edges.c
@@ -287,7 +287,7 @@ void edges_int_free(edges_int_t **edges) {
 	}
 }
 
-size_t edges_int_index_linear(edges_int_t const *edges, long long const value) {
+int edges_int_index_linear(edges_int_t const *edges, long long const value) {
 	return((value-edges->limits.lower)*edges->limits.bins / 
 		(edges->limits.upper-edges->limits.lower));
 }
